+++ src/Resource/AccountMessagingEmail.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace DigitalCz\DigiSign\Resource;
+
+class AccountMessagingEmail
+{
+    /** @var string */
+    public $type;
+
+    /** @var string[] */
+    public $subject;
+}