+++ src/Resource/EnvelopeRecipientIdentifications.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace DigitalCz\DigiSign\Resource;
+
+class EnvelopeRecipientIdentifications extends BaseResource
+{
+    /** @var Collection<EnvelopeRecipientIdentification> */
+    public $identifications;
+}